Final touches

+++ app/src/main/java/ba/kenan/myhabits/presentation/viewmodels/home/HomeViewModel.kt
@@ -3,10 +3,12 @@ package ba.kenan.myhabits.presentation.viewmodels.home
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import ba.kenan.myhabits.domain.model.Habit
+import ba.kenan.myhabits.domain.network.NetworkStatusProvider
 import ba.kenan.myhabits.domain.repository.HabitRepository
 import ba.kenan.myhabits.presentation.utils.SnackbarController
 import ba.kenan.myhabits.presentation.utils.SnackbarEvent
 import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.firestore.Source
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -19,9 +21,11 @@ import kotlin.coroutines.cancellation.CancellationException
 @HiltViewModel
 class HomeViewModel @Inject constructor(
     private val repository: HabitRepository,
-    private val snackbarController: SnackbarController
+    private val snackbarController: SnackbarController,
+    private val networkStatusProvider: NetworkStatusProvider
 ) : ViewModel() {
 
+    private val source = if (networkStatusProvider.isOffline()) Source.CACHE else Source.DEFAULT
     private val _uiState = MutableStateFlow<HomeUiState>(HomeUiState.Loading)
     val uiState: StateFlow<HomeUiState> = _uiState
 
@@ -29,7 +33,7 @@ class HomeViewModel @Inject constructor(
         viewModelScope.launch {
             _uiState.value = HomeUiState.Loading
             val uid = FirebaseAuth.getInstance().currentUser?.uid ?: return@launch
-            val result = repository.getHabitsForUser(uid)
+            val result = repository.getHabitsForUser(uid, source)
             _uiState.value = result.fold(
                 onSuccess = { HomeUiState.Success(it) },
                 onFailure = {
@@ -51,6 +55,13 @@ class HomeViewModel @Inject constructor(
         frequencyDays: List<Int>
     ) {
         viewModelScope.launch {
+            if (networkStatusProvider.isOffline()) {
+                snackbarController.sendEvent(
+                    SnackbarEvent("Can't add habits without an internet connection")
+                )
+                _uiState.value = HomeUiState.Error(Throwable(message = "No internet connection"))
+                return@launch
+            }
             val uid = FirebaseAuth.getInstance().currentUser?.uid ?: return@launch
             val result = repository.addHabit(uid, name, tags, frequencyType, frequencyDays)
             result.fold(
@@ -64,7 +75,7 @@ class HomeViewModel @Inject constructor(
                     snackbarController.sendEvent(
                         SnackbarEvent(it.message ?: "Couldn't add habit.")
                     )
-                    HomeUiState.Error(it)
+                    _uiState.value = HomeUiState.Error(it)
                 }
             )
         }
@@ -72,6 +83,12 @@ class HomeViewModel @Inject constructor(
 
     fun updateHabit(habit: Habit) {
         viewModelScope.launch {
+            if (networkStatusProvider.isOffline()) {
+                snackbarController.sendEvent(
+                    SnackbarEvent("Can't update habits without an internet connection")
+                )
+                return@launch
+            }
             val uid = FirebaseAuth.getInstance().currentUser?.uid ?: return@launch
             val result = repository.updateHabit(
                 userId = uid,
@@ -93,7 +110,7 @@ class HomeViewModel @Inject constructor(
                     snackbarController.sendEvent(
                         SnackbarEvent(it.message ?: "Couldn't update habit.")
                     )
-                    HomeUiState.Error(it)
+                    _uiState.value = HomeUiState.Error(it)
                 }
             )
         }
@@ -101,6 +118,12 @@ class HomeViewModel @Inject constructor(
 
     fun deleteHabit(habitId: String) {
         viewModelScope.launch {
+            if (networkStatusProvider.isOffline()) {
+                snackbarController.sendEvent(
+                    SnackbarEvent("Can't delete habits without an internet connection")
+                )
+                return@launch
+            }
             val uid = FirebaseAuth.getInstance().currentUser?.uid ?: return@launch
             val result = repository.deleteHabit(uid, habitId)
             result.fold(
@@ -122,12 +145,16 @@ class HomeViewModel @Inject constructor(
 
     fun markHabitAsCompleted(habitId: String) {
         viewModelScope.launch {
+            if (networkStatusProvider.isOffline()) {
+                snackbarController.sendEvent(
+                    SnackbarEvent("Can't complete habits without an internet connection")
+                )
+                return@launch
+            }
             val userId = FirebaseAuth.getInstance().currentUser?.uid ?: return@launch
             val date = LocalDate.now().format(DateTimeFormatter.ISO_DATE)
-
             repository.markHabitCompleted(userId, habitId, date)
-
-            val result = repository.getHabitsForUser(userId)
+            val result = repository.getHabitsForUser(userId, source)
             _uiState.value = result.fold(
                 onSuccess = {
                     snackbarController.sendEvent(
